# src/server/config/swagger.json
{
	"components": {
...
		"schemas": {
			"Guestbook": {
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"type": "string"
					},
					"listName": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"address",
					"listName",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FlattenMaps_T_": {
				"properties": {},
				"type": "object"
			},
			"Document_any.any.any_": {
				"$ref": "#/components/schemas/FlattenMaps_T_",
				"description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
			},
			"Item": {
				"properties": {
					"name": {
						"type": "string"
					},
					"rng": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IObjectWithTypegooseFunction": {
				"description": "Interface describing a Object that has a \"typegooseName\" Function",
				"properties": {},
				"type": "object",
				"additionalProperties": false
			},
			"ObjectId": {
				"type": "string"
			},
...
			"mongoose.Types.ObjectId": {
				"type": "string"
			},
...
			"APIKey_Scopes": {
				"enum": [
					"read",
					"write"
				],
				"type": "string"
			},
			"APIKeyDoc": {
				"description": "APIKey document with _id field.",
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"type": "string",
						"description": "APIKeys are linked to users"
					},
					"name": {
						"type": "string",
						"description": "Every APIKey should have a unique name"
					},
					"jwt": {
						"type": "string",
						"description": "Store source JWT for authorizaiton"
					},
					"hash": {
						"type": "string",
						"description": "Store hash of the JWT"
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"description": "Date the API Key expires"
					},
					"scopes": {
						"items": {
							"$ref": "#/components/schemas/APIKey_Scopes"
						},
						"type": "array",
						"description": "Scopes of the API Key relative to the User account"
					},
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"name",
					"expiresAt",
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api-key": {
				"type": "apiKey",
				"name": "token",
				"in": "query"
			},
			"siwe": {
				"type": "apiKey",
				"name": "authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Express OpenAPI Server",
		"description": "An example node app using OpenAPI - Express - Typescript. [Sign In With Ethereum](./../auth)",
		"contact": {
			"name": "@tyzoo",
			"url": "https://github.com/tyzoo",
			"email": "tyler.russell.design@gmail.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/guestbook": {
			"get": {
				"operationId": "FindAllitems",
				"responses": {
					"200": {
						"description": "Array of Items",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"$ref": "#/components/schemas/Guestbook"
											},
											{
												"properties": {
													"_id": {
														"type": "string"
													}
												},
												"required": [
													"_id"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all items",
				"summary": "Get all items",
				"tags": [
					"Guestbook"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateItem",
				"responses": {
					"200": {
						"description": "New Item",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Guestbook"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"description": "Create an Item",
				"summary": "Create an Item",
				"tags": [
					"Guestbook"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Provide new document contents",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"message": {
										"type": "string"
									},
									"address": {
										"type": "string"
									}
								},
								"required": [
									"message",
									"address"
								],
								"type": "object",
								"description": "Provide new document contents"
							}
						}
					}
				}
			}
		},
		"/crud": {
			"get": {
				"operationId": "FindAllItems",
				"responses": {
					"200": {
						"description": "Array of Items",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"allOf": [
											{
												"$ref": "#/components/schemas/Document_any.any.any_"
											},
											{
												"$ref": "#/components/schemas/Item"
											},
											{
												"$ref": "#/components/schemas/IObjectWithTypegooseFunction"
											},
											{
												"properties": {
													"_id": {
														"$ref": "#/components/schemas/ObjectId"
													}
												},
												"required": [
													"_id"
												],
												"type": "object"
											}
										]
									},
									"type": "array"
								},
								"examples": {
									"Successful Response": {
										"value": [
											{
												"_id": "62f05f918c8a2e1d6608dfd2",
												"name": "Item 4",
												"rng": 0.9992763155631645
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Get all items",
				"summary": "Get all items (With Pagination, Sorting, Filtering)",
				"tags": [
					"CRUD"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateItem",
				"responses": {
					"200": {
						"description": "New Item",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Item"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								},
								"examples": {
									"Successful Response": {
										"value": {
											"_id": "62f05f918c8a2e1d6608dfd2",
											"name": "Item 4",
											"rng": 0.9992763155631645
										}
									}
								}
							}
						}
					}
				},
				"description": "Create an Item",
				"summary": "Create an Item",
				"tags": [
					"CRUD"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Provide new document contents",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object",
								"description": "Provide new document contents"
							}
						}
					}
				}
			}
		},
		"/crud/{itemId}": {
			"get": {
				"operationId": "FindItem",
				"responses": {
					"200": {
						"description": "Item",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Item"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									],
									"nullable": true
								},
								"examples": {
									"Successful Response": {
										"value": {
											"_id": "62f05f918c8a2e1d6608dfd2",
											"name": "Item 4",
											"rng": 0.9992763155631645
										}
									}
								}
							}
						}
					}
				},
				"description": "Get Item by ID",
				"summary": "Get Item By ID",
				"tags": [
					"CRUD"
				],
				"security": [],
				"parameters": [
					{
						"description": "Document MongoID",
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/mongoose.Types.ObjectId"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateItem",
				"responses": {
					"200": {
						"description": "Updated Item",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Item"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								},
								"examples": {
									"Successful Response": {
										"value": {
											"_id": "62f05f918c8a2e1d6608dfd2",
											"name": "Item 4",
											"rng": 0.9992763155631645
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Unauthorized request"
										}
									}
								}
							}
						}
					}
				},
				"description": "Update an Item",
				"summary": "Update an Item (User Only)",
				"tags": [
					"CRUD"
				],
				"security": [
					{
						"jwt": [
							"user"
						]
					}
				],
				"parameters": [
					{
						"description": "Document MongoID",
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/mongoose.Types.ObjectId"
						}
					}
				],
				"requestBody": {
					"description": "Content to update to",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"name": {
										"type": "string"
									}
								},
								"required": [
									"name"
								],
								"type": "object",
								"description": "Content to update to"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteItem",
				"responses": {
					"200": {
						"description": "Deleted Item",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Item"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								},
								"examples": {
									"Successful Response": {
										"value": {
											"_id": "62f05f918c8a2e1d6608dfd2",
											"name": "Item 4",
											"rng": 0.9992763155631645
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Unauthorized request"
										}
									}
								}
							}
						}
					}
				},
				"description": "Delete an Item",
				"summary": "Delete an Item  (Admin Only)",
				"tags": [
					"CRUD"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Document MongoID",
						"in": "path",
						"name": "itemId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/mongoose.Types.ObjectId"
						}
					}
				]
			}
		},
...
		"/api-keys": {
			"get": {
				"operationId": "FindAllitems",
				"responses": {
					"200": {
						"description": "- An array of APIKeys.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/APIKeyDoc"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Unauthorized request"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieves all APIKeys.",
				"summary": "Retrieves all APIKeys.",
				"tags": [
					"APIKeys"
				],
				"security": [
					{
						"siwe": [
							"user"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateApiKey",
				"responses": {
					"200": {
						"description": "- The new APIKey.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIKeyDoc"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Unauthorized request"
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a new APIKey.",
				"summary": "Creates a new APIKey.",
				"tags": [
					"APIKeys"
				],
				"security": [
					{
						"siwe": [
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- Request body.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"tokenName": {
										"type": "string"
									},
									"scopes": {
										"items": {
											"$ref": "#/components/schemas/APIKey_Scopes"
										},
										"type": "array"
									},
									"expiresInDays": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"tokenName",
									"scopes",
									"expiresInDays"
								],
								"type": "object",
								"description": "- Request body."
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}